requirements.txt
# AI Media Translation & Publishing Agent dependencies

# Core Inference / AI Models
faster-whisper==1.0.3        # Stage 2: Fast local ASR
pyannote.audio>=3.3.1        # Stage 2: Speaker Diarization
TTS==0.22.0                  # Stage 4: Coqui XTTSv2 Voice Cloning
deep-translator==1.11.4      # Stage 3: Free Translation (Google Translate fallback)

# Audio Processing
pydub==0.25.1                # Stage 5: Audio mixing, EQ
pyloudnorm==0.1.1            # Stage 5: LUFS normalization
soundfile==0.12.1            # Audio file I/O

# Video Processing & Subtitles
ffmpeg-python==0.2.0         # Stages 1, 6-8: Video processing, audio extraction, burn-in
opencv-python==4.9.0.80      # Stage 9: Video properties validation

# Web UI / Application
streamlit==1.32.2            # Simple frontend UI
python-dotenv==1.0.1         # Environment variable management

# Utilities
numpy
scipy
Requests
tqdm
google-generativeai
